core/requirements.txt
# ── Core ──
fastapi>=0.110,<1.0
uvicorn>=0.27,<1.0
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0
httpx>=0.27,<1.0
sse-starlette>=2.0,<3.0
python-multipart>=0.0.9,<1.0

# ── Gemini (Built-in LLM) ──
google-generativeai>=0.8

# ── MCP SDK (Plugin Infrastructure) ──
mcp>=1.0

# ── Legacy LLM (Phase 5 で MCP 移行予定) ──
openai>=1.12,<2.0
langgraph>=0.2
langchain-core>=0.3
langchain>=0.3

# ── Data Processing ──
openpyxl>=3.1,<4.0

# ── OpenVINO (Optional) ──
openvino>=2024.4.0
openvino-genai>=2024.4.0
openvino-tokenizers>=2024.4.0

# ── YAML (Skill definitions) ──
pyyaml>=6.0
